refactor(#792): extract per-segment audio DSP into runtime_process_segment.rs (sweep)
runtime.rs 888 -> 480 (leaves the debt list). The two callback entry points
(process_input_f32 / process_output_f32[_mixed]) stay; the per-segment hot-path
DSP moves to runtime_process_segment.rs:

  - process_single_segment (+ SegmentFeed) — reads the segment's frames, runs
    its block chain, taps, fade-in, spillover, and mixes into the route buffer
  - process_audio_block / apply_block_processor — per-block fade + panic guard
  - mix_outgoing_tail — #454-T5 spillover tail
  - downcast_panic_message

Pure mechanical move — no logic change, RT-safe contract unchanged (module
boundaries are compile-time only; zero runtime cost). process_input_f32 calls
the now-pub(crate) process_single_segment; the test-only helpers the suite
exercises via super:: are re-exported under #[cfg(test)].

Verified: engine suite green (alloc-invariant / golden / deadline / volume /
isolation invariants pass), all workspace test targets compile clean, zero
warnings.
